requirements.txt
pytest==7.1.3
python-dotenv==0.21.0
PyVimeo
selenium==4.5.0
vimeo
webdriver_manager==3.8.3
pywin32
